[Navigation] - Fixes the toolbar title truncation issue
