+++ main/src/main/java/edu/artic/main/MainActivity.kt
@@ -1,10 +1,7 @@
 package edu.artic.main
 
 
-import android.app.Activity
-import android.graphics.Color
 import android.os.Bundle
-import android.view.WindowManager
 import edu.artic.base.utils.disableShiftMode
 import edu.artic.viewmodel.BaseViewModelActivity
 import kotlinx.android.synthetic.main.activity_main.*
@@ -25,20 +22,7 @@ class MainActivity : BaseViewModelActivity<MainViewModel>() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setWindowFlag(this, WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS, false)
-        window.statusBarColor = Color.TRANSPARENT
         bottomNavigation.disableShiftMode(R.color.menu_color_list)
-    }
-
-    private fun setWindowFlag(activity: Activity, bits: Int, on: Boolean) {
-        val win = activity.window
-        val winParams = win.attributes
-        if (on) {
-            winParams.flags = winParams.flags or bits
-        } else {
-            winParams.flags = winParams.flags and bits.inv()
-        }
-        win.attributes = winParams
     }
 
 }
