+++ viewmodel/src/main/kotlin/edu/artic/viewmodel/BaseViewModelFragment.kt
@@ -1,8 +1,12 @@
 package edu.artic.viewmodel
 
+import android.app.Activity
 import android.arch.lifecycle.ViewModelProvider
+import android.graphics.Color
 import android.os.Bundle
+import android.support.v4.content.ContextCompat
 import android.view.View
+import android.view.WindowManager
 import edu.artic.ui.BaseActivity
 import edu.artic.ui.BaseFragment
 import javax.inject.Inject
@@ -15,8 +19,6 @@ import kotlin.reflect.KClass
 abstract class BaseViewModelFragment<TViewModel : BaseViewModel> : BaseFragment() {
 
     protected abstract val viewModelClass: KClass<TViewModel>
-
-
     /**
      * @return True by default if we use fragment for view model provider. Otherwise we use the [BaseActivity]
      * * as where the [TViewModel] lives.
@@ -47,7 +49,7 @@ abstract class BaseViewModelFragment<TViewModel : BaseViewModel> : BaseFragment(
     }
 
     val viewModel: TViewModel by viewModelLazy
-    var isViewJustCreated : Boolean = true
+    var isViewJustCreated: Boolean = true
     protected open fun getViewModelForClass(viewModelProvider: ViewModelProvider,
                                             kClass: KClass<TViewModel>): TViewModel =
             viewModelProvider.get(kClass.java)
@@ -61,11 +63,36 @@ abstract class BaseViewModelFragment<TViewModel : BaseViewModel> : BaseFragment(
             viewModel.register(this)
             onRegisterViewModel(viewModel)
         }
+        if (hasTransparentStatusBar()) {
+            activity?.let {
+                setWindowFlag(it, WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS, false)
+                it.window.statusBarColor = Color.TRANSPARENT
+            }
+        } else {
+            activity?.let {
+                it.window?.statusBarColor = ContextCompat.getColor(it, R.color.colorPrimary)
+            }
+        }
+    }
+
+
+
+    protected open fun hasTransparentStatusBar(): Boolean = false
+
+    private fun setWindowFlag(activity: Activity, bits: Int, on: Boolean) {
+        val win = activity.window
+        val winParams = win.attributes
+        if (on) {
+            winParams.flags = winParams.flags or bits
+        } else {
+            winParams.flags = winParams.flags and bits.inv()
+        }
+        win.attributes = winParams
     }
 
     override fun onResume() {
         super.onResume()
-        if(isViewJustCreated) {
+        if (isViewJustCreated) {
             setupBindings(viewModel)
         }
         setupNavigationBindings(viewModel)
